+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated default style set to cell from Height to Min-Height to support large text content.
+
 ## 5.4.0 - (January 12, 2024)
 
 * Added