+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added fromWorkListDataGrid a private prop that gets passed down from terra-data-grid.
+
 * Changed
   * Replaced cell `height` property with `min-height` property to support large text content inside cell.